fix: Separate metadata export and SignIn component to resolve client/server component conflict
