+++ src/app/signin/metadata.ts
@@ -1,0 +1,5 @@
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'Sign In',
+};
